Fix a bug in start/stop command codegen & add show/hide commands.
We were generating the start/stop instructions with a 0 parameter, when it should have been BACK_OF_STACK. In some cases this turned out to be the same, but in complex expressions this isn't the same stack location. Also renamed the start/stop commands/instructions to no longer mention movies.