Working on porting the document structure (stacks, cards, parts etc.) to C++.
In the process, it should become possible to asynchronously load stacks over the network, card-by-card, and to backwards-compatibly save new properties as user properties. Also, anyone who wants to port this to another platform should have less ObjC-code they need to re-create in a more common language.
None of this is currently hooked up, so you shouldn't see any effect of this.